refactor: 예외 클래스 패키지 이동 및 전역 처리 클래스 리팩토링
