--- src/main/java/com/nhnacademy/bookstoreorderapi/common/exception/OrderNotFoundException.java
@@ -1,7 +1,0 @@
-package com.nhnacademy.bookstoreorderapi.common.exception;
-
-public class OrderNotFoundException extends RuntimeException {
-    public OrderNotFoundException(String orderId) {
-        super("주문을 찾을 수 없습니다: orderId=" + orderId);
-    }
-}
