+++ src/main/java/com/nhnacademy/bookstoreorderapi/order/exception/OrderNotFoundException.java
@@ -1,0 +1,7 @@
+package com.nhnacademy.bookstoreorderapi.order.exception;
+
+public class OrderNotFoundException extends RuntimeException {
+    public OrderNotFoundException(String orderNumber) {
+        super("주문을 찾을 수 없습니다: orderNumber=" + orderNumber);
+    }
+}
